requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
crispy-bootstrap4==2022.1
cryptography==41.0.4
Deprecated==1.2.14
Django==4.2.5
django-cors-headers==4.2.0
django-crispy-forms==2.0
django-environ==0.11.2
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
idna==3.4
jwcrypto==1.5.0
oauthlib==3.2.2
psycopg2-binary==2.9.9
pycparser==2.21
python-http-client==3.3.7
pytz==2023.3.post1
requests==2.31.0
sendgrid==6.10.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
urllib3==2.0.6
wrapt==1.15.0
gunicorn==20.1.0